fix: scale down background so it works on firefox
I noticed that on firefox it would fail to load the background but
work on Chrome. My hypothesis is that Firefox has a lower max
texture size than Chrome after looking around. A way to resolve this
is to scale down the image (I noticed we can scale down the raw image
cleanly by 25x) and scale it back up as a _sprite_. This works on
both Chrome and Firefox now.

One thing we had to do to make this work though is also disable image
optimizations for the background - otherwise, the background looks awful
due to compression or webp optimizations. I noticed the generated webp
file is basically the same size as the raw png file though, so I don't
think this is a problem.